# Move checkpointing parallelism into `TaskExecutor` class, use that class for parallel `union_by_name`