--- a/src/parallel/CMakeLists.txt
+++ b/src/parallel/CMakeLists.txt
@@ -13,6 +13,7 @@ add_library_unity(
   pipeline_executor.cpp
   pipeline_finish_event.cpp
   pipeline_initialize_event.cpp
+  task_executor.cpp
   task_scheduler.cpp
   thread_context.cpp)
 set(ALL_OBJECT_FILES